backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pdfplumber==0.10.3
python-multipart==0.0.6
pymongo==4.6.1
pydantic==2.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
google-generativeai==0.3.2
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
pypdfium2==4.25.0
